Added route to get user by id

+++ src/api/users/paramValidation.js
@@ -67,6 +67,5 @@ export const getUserById = [
     .exists()
     .withMessage('ID is missing from params')
     .isUUID()
-    .withMessage('ID should be a uuid')
-    .toInt(),
+    .withMessage('ID should be a uuid'),
 ];
